Filter personalisation before passing to notify.
This should hopefully stop any markdown being passed through and
applied from citizen emails. We should be applying similar approaches
to emails sent from the back-office, but those are only sent to the
applicant themselves so have a smaller impact than these which are
to other citizens with gov.uk branding etc.

The emails sent will look awful but that's probably better than being
a vehicle for phishing etc.